Scroll behavior component (#1107)
* Naive first step to refactor out scroll logic into separate component

* Move a bunch of event logic to `scroll_component` from `Scroll`

* Add method to coordinate painting with `scroll_component`

* Move a bunch of lifecycle logic to `scroll_component` from `Scroll`

* Documentation pass on `scroll_component`

* Rename `Scroll` to `AbsoluteScroll` & add scroll functionality to `List`

* Add ability to change layout axis for `List` with builder fn

* Scroll behavior component changelog

* Pass `BarHoveredState` by value and add impl Default for ScrollComponent

* `ScrollComponent` review changes

* Add some missing information to `ScrollComponent` documentation

* `ScrollContainer` review take 2

* Changes from #1057

* ScrollComponent clippy

* Remove List changes (updated to master) & fix Scroll regression

* Add missing doc comments

* Be a bit more careful about event consuming

* Fix dup changelog entry & doc comment misspelling
