+++ AUTHORS
@@ -11,4 +11,5 @@ Kaiyin Zhong
 Kaur Kuut
 Leopold Luley
 Andrey Kabylin
+Garrett Risley
 Robert Wittams
